# arm64: optimized utf32_to_utf16 for regular non-surrogate case