I observed that the routine does a narrow from 32 to 16 and uses 8byte SIMD.
This PR processes 32 bytes in the loop instead of 16 and upgrades the instructions to process ~2x compared to previous one.

Further surrogate loop optimization can be done separately.

Before:
```
convert_utf32_to_utf16be+arm64, input size: 260, iterations: 30000, dataset: testutf32.txt
  15.182 GB/s (14.9 %)    3.737 Gc/s     4.06 byte/char     17.1 ns
WARNING: Measurements are noisy, try increasing iteration count (-I).
convert_utf32_to_utf16be_with_errors+arm64, input size: 260, iterations: 30000, dataset: testutf32.txt
  13.520 GB/s (9.8 %)    3.328 Gc/s     4.06 byte/char     19.2 ns
convert_utf32_to_utf16le+arm64, input size: 260, iterations: 30000, dataset: testutf32.txt
  17.660 GB/s (17.4 %)    4.347 Gc/s     4.06 byte/char     14.7 ns
WARNING: Measurements are noisy, try increasing iteration count (-I).
convert_utf32_to_utf16le_with_errors+arm64, input size: 260, iterations: 30000, dataset: testutf32.txt
  14.423 GB/s (18.0 %)    3.550 Gc/s     4.06 byte/char     18.0 ns
WARNING: Measurements are noisy, try increasing iteration count (-I).
```
After:
```
convert_utf32_to_utf16be+arm64, input size: 260, iterations: 30000, dataset: testutf32.txt
  27.040 GB/s (16.8 %)    6.656 Gc/s     4.06 byte/char      9.6 ns
WARNING: Measurements are noisy, try increasing iteration count (-I).
convert_utf32_to_utf16be_with_errors+arm64, input size: 260, iterations: 30000, dataset: testutf32.txt
  23.708 GB/s (15.6 %)    5.836 Gc/s     4.06 byte/char     11.0 ns
WARNING: Measurements are noisy, try increasing iteration count (-I).
convert_utf32_to_utf16le+arm64, input size: 260, iterations: 30000, dataset: testutf32.txt
  27.040 GB/s (12.5 %)    6.656 Gc/s     4.06 byte/char      9.6 ns
WARNING: Measurements are noisy, try increasing iteration count (-I).
convert_utf32_to_utf16le_with_errors+arm64, input size: 260, iterations: 30000, dataset: testutf32.txt
  24.038 GB/s (10.6 %)    5.917 Gc/s     4.06 byte/char     10.8 ns
WARNING: Measurements are noisy, try increasing iteration count (-I).
```